Improve SourceId to allow for URLs (#80)
- `SourceId` now only stores a `String`, not a `Vec<String>`
- I first wrote this as an enum with:

```rust
enum SourceId {
    Empty,
    Repl,
    Path(PathBuf),
    Url(Url)
}
```

But `PathBuf` comes from `std` and `Url` is a sizeable dependency, so
this way the user can decide how they want to use the inside of the
SourceId.

This is for https://github.com/ahdinosaur/avo/pull/4, where I've started
using `rimu` to implement a Salt Stack / NixOS alternative.